% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Roda o aplicativo CultivaR}
\usage{
run_app(db_path = NULL, ...)
}
\arguments{
\item{db_path}{Caminho para o arquivo SQLite. Default: DB_PATH_DEV
(resolvido relativo ao diretorio do projeto).}

\item{...}{Argumentos passados para shiny::shinyApp().}
}
\description{
Roda o aplicativo CultivaR
}
